fix: fix the touch board issue
Previously, when computer users scrolled up and down in the terminal
area using two fingers on a touchpad, reaching the top or bottom would
inadvertently trigger changes to the terminal's initialChildSize,
causing all expand-related functionalities to fail. By setting the
ListView's controller to null on desktop and web platforms, this issue
can be fixed. Two-finger scrolling up and down still works when the
focus is on the terminal, while page scrolling remains functional.

The keypoint is to seperate the DraggableScrollableSheet controller and
ListView controller. Once they are different instance. The problem
solved.

Signed-off-by: Dennis Liu <dennis48161025@gmail.com>